refactor(relayer): Factor out common listener utils
This is a pre-emptive change made to make it easier to reuse common
components in alternative listener implementations. Examples for a
viem-based listener, as well as an eventual Solana listener.
